fix(sdk): restore error reporting after esbuild@0.14.4 broke previous behavior
